added Swagger config and help page
