+++ src/main/java/com/example/sortener/Application.java
@@ -2,6 +2,7 @@ package com.example.sortener;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+//import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 
 @SpringBootApplication
 public class Application {
