--- src/main/java/com/example/sortener/Constants/AppConstants.java
@@ -1,11 +1,0 @@
-package com.example.sortener.Constants;
-
-public class AppConstants {
-
-    public static final String OPENED_ACCOUNT = "Your account is opened.";
-    public static final String EXIST_ACCOUNT = "Your account already exist.";
-    public static final String APP_LINK = "http://localhost:8080/";
-
-    public static final String CHARACTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789~`!@#$%^&*()-_=+[{]}\\|;:\'\",<.>/?";
-
-}
