+++ src/main/java/com/example/sortener/assembler/AccountAssembler.java
@@ -4,8 +4,8 @@ import com.example.sortener.dto.account.ResponseAccountDto;
 import com.example.sortener.entity.Account;
 import org.springframework.stereotype.Component;
 
-import static com.example.sortener.Constants.AppConstants.EXIST_ACCOUNT;
-import static com.example.sortener.Constants.AppConstants.OPENED_ACCOUNT;
+import static com.example.sortener.constants.AppConstants.EXIST_ACCOUNT;
+import static com.example.sortener.constants.AppConstants.OPENED_ACCOUNT;
 import static java.util.Objects.isNull;
 
 @Component
