+++ src/main/java/com/example/sortener/constants/AppConstants.java
@@ -1,0 +1,17 @@
+package com.example.sortener.constants;
+
+public class AppConstants {
+
+    public static final String OPENED_ACCOUNT = "Your account is opened.";
+    public static final String EXIST_ACCOUNT = "Your account already exist.";
+    public static final String APP_LINK = "http://localhost:8080/";
+
+    public static final String CHARACTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789~`!@#$%^&*()-_=+[{]}\\|;:\'\",<.>/?";
+
+    public static final String HELP_PAGE = "INSTALLATION \n \n\n 1. Run mvc clean package to create executable JAR file \n\n " +
+            "2. JAR File is contained in root/target/url/Shortener-0.0.1-SNAPSHOT.jar \n\n " +
+            "3. Find the file through any kind of cmd/terminal and run command java-jar-urlShortener-0.0.1-UrlShortener.jar \n\n " +
+            "4. This will start up whole application \n\n " +
+            "5. Add row (127.0.0.1) \n\n\n 1. Swagger je tam\n\n 2. Treba ti ucet ";
+
+}
