+++ src/main/java/com/example/sortener/controller/HelpController.java
@@ -1,17 +1,14 @@
 package com.example.sortener.controller;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 
-import javax.servlet.http.HttpServletRequest;
-import java.util.Map;
-
-import static java.lang.System.currentTimeMillis;
+import static com.example.sortener.constants.AppConstants.HELP_PAGE;
 import static org.springframework.http.ResponseEntity.ok;
 
 @RestController
@@ -20,13 +17,9 @@ import static org.springframework.http.ResponseEntity.ok;
 @RequiredArgsConstructor
 public class HelpController {
 
-    @GetMapping("help")
-    public ResponseEntity<Map<String, Integer>> getStatistics(HttpServletRequest httpServletRequest, @PathVariable String accountId) {
-        validator.authenticate(httpServletRequest);
-        long runTime = currentTimeMillis();
-        var statistics = accountService.getStatistics(accountId);
-        logger.debug("Getting statistics for user: {} took {} millis", accountId, currentTimeMillis() - runTime);
-        return ok(statistics);
+    @GetMapping(path = "help", produces = MediaType.TEXT_PLAIN_VALUE)
+    public ResponseEntity<String> getHelp() {
+        return ok(HELP_PAGE);
     }
 
 }
